refactor: 테스트 유연성을 위한 의존성 주입 도입
